feat(posts): 記事一覧をページネーションから単一ページ表示に変更
SSGで /posts/[page] のURL付きページネーションを使うと、新記事追加時に
canonicalが壊れる問題があったため、/posts 単一URLで全記事を1ページに
表示する構成に変更。
