+++ src/constants/index.ts
@@ -1,11 +1,9 @@
 import type { Entry } from '@/types/entry'
 
-export const PER_PAGE = 12
-
 export const ENTRIES: Entry[] = [
   { href: '/', content: 'home' },
   { href: '/about/', content: 'about' },
-  { href: '/posts/1/', content: 'post' },
+  { href: '/posts/', content: 'post' },
 ]
 
 export const OGP_FONT = 'M PLUS 1p'
